improvement(k8s): syncs keep running after Garden process exits
Also, we now only run one daemon per Garden project at any time, with a single
data directory. None of that temp directory business, symlinks and whatnot.

Further, we get updates from Mutagen instantly via a single
`mutagen sync monitor` process instead of execing a `mutagen sync list` process
every 2 seconds.

A side-effect of that is that our logging is maybe a touch noisy. I left a TODO
to address that.
